requirements.txt
# Core Data Science
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.3.0
xgboost>=1.7.6

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# MLOps & Tracking
mlflow>=2.6.0
dvc>=3.32.0
evidently>=0.4.13

# API & Web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
pydantic>=2.0.0

# Streamlit Extras (for 3D GUI)
streamlit-extras>=0.3.0
streamlit-lottie>=0.0.5
streamlit-option-menu>=0.3.6

# Testing & Quality
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
jupyter>=1.0.0
notebook>=7.0.6
joblib>=1.3.0

# For data download
requests>=2.31.0
kaggle>=1.6.12
